+++ src/test/java/io/zerogone/service/BlogSearchServiceTest.java
@@ -3,7 +3,6 @@ package io.zerogone.service;
 import io.zerogone.config.DatabaseConfiguration;
 import io.zerogone.config.WebConfiguration;
 import io.zerogone.exception.NotExistedDataException;
-import io.zerogone.model.CurrentUserInfo;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
@@ -17,8 +16,6 @@ import org.springframework.test.context.web.AnnotationConfigWebContextLoader;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.web.context.WebApplicationContext;
 
-import javax.persistence.NoResultException;
-
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = {WebConfiguration.class, DatabaseConfiguration.class}, loader = AnnotationConfigWebContextLoader.class)
 @WebAppConfiguration
@@ -31,13 +28,6 @@ public class BlogSearchServiceTest {
     @Before
     public void setUp() {
         blogSearchService = webApplicationContext.getBean(BlogSearchService.class);
-    }
-
-    @Test
-    public void getBlogsThatUserBelongTo() {
-        CurrentUserInfo currentUserInfo = new CurrentUserInfo();
-        currentUserInfo.setId(1);
-        Assert.assertNotNull(blogSearchService.getBlogsThatUserBelongTo(currentUserInfo));
     }
 
     @Test
